requirements.txt
pandas
plotly
snowflake-snowpark-python[pandas]
streamlit
